soundmanager틀 작성
