+++ Assets/Resources/Sounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2a2980410377e5e44be056ee06ff7b7a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
