+++ Assets/Resources/Sounds/(LOOP-READY) Track 1 - Intro.wav.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 91a0ac7f6ba22e6469afb3f54fdcccac
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
