+++ Assets/Script/DB/Manager.cs
@@ -1,31 +1,33 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+//using System.Collections;
+//using System.Collections.Generic;
+//using UnityEngine;
 
-public class Manager : MonoBehaviour
-{
-
-    static DBConnectionInfo dBConnectionInf = new DBConnectionInfo
-    {
-        ipAddress = "localhost",
-        user = "root",
-        password = "",
-        dbName = "mydb"
-    };
+//public class Manager : MonoBehaviour
+//{
     
-    DBManager DBManager = new DBManager(dBConnectionInf);
-    // Start is called before the first frame update
-    void Start()
-    {
-        Debug.Log(DBManager.GetMoney());
-        List<SkillDataStruct> List = DBManager.GetSkillData();
-        HighScore highScore;
-        DBManager.GetRecordHighScore(out highScore);
-    }
+//    DBConnectionInfo dBConnectionInf = new DBConnectionInfo
+//    {
+//        ipAddress = "localhost",
+//        user = "root",
+//        password = "",
+//        dbName = "mydb"
+//    };
 
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
+//    DBManager DBManager;
+
+//    void Start()
+//    {
+//        DBManager DBManager = new DBManager(dBConnectionInf);
+//        Debug.Log(DBManager.GetMoney());
+//        List<SkillDataStruct> List = DBManager.GetSkillData();
+//        HighScore highScore;
+//        DBManager.GetRecordHighScore(out highScore);
+//    }
+
+//    // Update is called once per frame
+//    void Update()
+//    {
+//        DBManager = 
+//    }
+    
+//}
