+++ Assets/Script/SoundTest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99f09a8378d947947a9b7ea1a8a19cc5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
