+++ Assets/Script/SoundTest/BGMTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4907c59b8250cf042b1cf9c8c9c88d8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
