+++ Assets/Script/SoundTest/SoundManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36556384ab24d094aab6706d57ca5eb0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
